Fix incorrect variable expansion when parsing PalServer.sh start options
The way the script is currently implemented, it passes all of the start options as one quoted string to `PalServer.sh`. This results in Palworld itself setting all of the options (including spaces) as a single argument, leading to the options not being properly applied.

This PR changes `START_OPTIONS` to a Bash array, where the required options are added incrementally as required to the array, and when ready, the whole array is expanded as individual arguments to `PalServer.sh`.

Fixes #186.